DEV-923: Moved algorithms to shared folder
