+++ src/com/xqmsg/sdk/shared/CommunicationsEnum.ts
@@ -1,0 +1,8 @@
+/** An enum denoting the various communication types the user can encrypt */
+export enum CommunicationsEnum {
+  UNKNOWN = "unknown",
+  FILE = "file",
+  EMAIL = "email",
+  CHAT = "chat",
+  FORM = "form",
+}
