+++ src/com/xqmsg/sdk/v2/CommunicationsEnum.ts
@@ -1,8 +1,1 @@
-/** An enum denoting the various communication types the user can encrypt */
-export enum CommunicationsEnum {
-  UNKNOWN = "unknown",
-  FILE = "file",
-  EMAIL = "email",
-  CHAT = "chat",
-  FORM = "form",
-}
+export { CommunicationsEnum } from "../shared/CommunicationsEnum";
